+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## [Unreleased]
 ## Added
 - Support adding custom events. (Check the tests/tests.rs file for an example.)
+- Support `path` property of a condition.
 ## Changed
 ## Removed